Added the Moodle's Teaching EPOS initial files over the 'svn/teaching/ine5424'.
https://epos.lisha.ufsc.br/svn/teaching/ine5424

+++ TeachingEpos/src/utility/random.cc
@@ -2,6 +2,6 @@
 
 __BEGIN_UTIL
 
-int Random::_seed;
+unsigned long int Random::_seed;
 
 __END_UTIL
